+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # v0.4.3
 
+## Features
+
+* `IntoIterator` for `A1` and `RangeForCell`
 * support for parsing quoted sheet names (i.e., `'My Spreadsheet'!A1`)
+
+## **Breaking Changes**
+
+* `A1Iterator` no longer has a lifetime
 
 # v0.4.2